+++ app/controllers/dwc_occurrences_controller.rb
@@ -1,0 +1,8 @@
+class DwcOccurrencesController < ApplicationController
+  include DataControllerConfiguration::ProjectDataControllerConfiguration
+
+  def metadata
+    @dwc_occurrences = DwcOccurrence.where(project_id: sessions_current_project_id)
+  end
+
+end